requirements.txt
asgiref==3.2.9
Django==3.0.7
django-cors-headers==3.4.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jsonapi==3.1.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
inflection==0.5.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1